+++ src/services/index.js
@@ -1,2 +1,3 @@
 export * from "./member";
 export * from "./auth";
+export * from "./product";
